fix(customer):
- fix currency service's pom.xml file.
- fix currency service's DB migration file.
- fix currency service's mapping logic.

+++ services/currency-service/src/main/resources/db/migration/V1__create_currancy_rate_table.sql
@@ -2,6 +2,6 @@ create table if not exists currency.rate (
     id uuid unique not null primary key default gen_random_uuid(),
     currency_from varchar(3) not null,
     currency_to varchar(3) not null,
-    rate bigint not null,
+    rate decimal not null,
     last_updated timestamp not null default current_timestamp
 );
